AtomsBuilder

Documentation for AtomsBuilder.

AtomsBuilder._flexible_system

AtomsBuilder.bulk

AtomsBuilder.rattle!

Base.repeat

AtomsBuilder._flexible_system — Method

Helper function to convert construct a FlexibleSystem from a list of positions, elements, cell matrix and pbc tuple

AtomsBuilder.bulk — Method

bulk(sym) : generates a FlexibleSystem unit cell for a bulk crystal structure. If sym is a chemical symbol then the phase and lattice constant are taken from a database that is consistent with ASE. If sym is one of [:fcc, :bcc, :diamond, :hcp] then one needs to specify the kwargs a or c to determine the lattice constants.

AtomsBuilder.rattle! — Method

rattle!(at, r::Float64) -> at

Randomly perturbs the atom positions within a ball of radius r. The perturbation is uniform in angular component, and uniform in radial component. (Note this is not the same as choosing them uniform in cartesian coordinates!).

Base.repeat — Method

repeat(at, n::NTuple{3})
repeat(at, n::Integer)

Takes a structure and repeats it n_j times into the j-th cell-vector direction. For example,

at = repeat(bulk(:C), (3,2,4))

creates 3 x 2 x 4 unit cells of carbon.

The same can be achieved by *:

at = bulk(:) * (3, 2, 4)